move box selection algo from BoxSelectors back to DefaultBoxSelector;
